Check return value before returning err
Calls to the C MagickWand API almost universally return a
MagickBooleanType value indicating whether the call succeeded or not.
Instead of checking that return value, most of our wrapper functions
would just return mw.GetLastError, which calls MagickGetException under
the hood. For cases when the call actually did not succeed, this worked
fine; the exception would be retrieved, no false positives. However,
MagickGetException also returns for non-fatal warning, which was causing
errors to be returned even when the API call completed without any
issue.

This commit changes all of the code using mw.GetLastError to start
checking the return value of the underlying API call before actually
fetching the exception.