Fix dereferencing / free memory issues with GetImageHistogram
An issue is that copies of the PixelWand objects are being made (and
returned). This causes C.DestroyPixelWand to be called twice on a
*C.PixelWand leading to a fatal error.

In addition there were occasions where numberColors != len([]PixelWand)
due to checks being simply if the pointer's value was not nil. This
caused a seg-fault when accessing [idx]PixelWand values where
idx >= numberColors, which can commonly happen when doing a for-each
loop.