im7: Add func GetImageProfileBytes (#320)
The C.MagickGetImageProfile() function returns a char array but it is
not only the name of the profile but the bytes of the actual ICC
profile. Add the function GetImageProfileBytes() to return []byte the
same way RemoveImageProfile() does it. Keep the existing function
GetImageProfile() to return a string to not break the existing API but
adjust it to call GetImageProfileBytes() internally.